Add import panel in flask example
As described in https://discourse.holoviz.org/t/error-when-running-the-example-to-embed-plot-in-flask-application/2785, when panel is embedded in a flask app, if panel is not imported, the following error appears. 

KeyError: “View model name ‘panel.models.location.Location’ not found”

this happens because bokeh does not know how to handle the foreign panel models.